Solve issue with exceptions being thrown in .map
Solved an issue where SolvingStrategies would throw a
PossibilityIntegrityViolatedException, which led to issues with
Java's Stream API that doesn't allow for checked exceptions to
be thrown.

The issue was "solved" by applying a somewhat smelly technique that
sneaks a checked exception past the compiler.
Advantage: The compiler requires the exception to be caught.